add optional pos parameter to start search from
This change is backward compatible.  It adds an additional int (or rather ssize_t types) pos parameter to ba.search,
and the search is started from that position on.  That significantly improves runtime for continuous searches in large
bitarrays.  In my specific use case of searching 1k distinct bit patterns 1M times in a 16MBit array, the time reduced from 80 seconds to 0.5 seconds.